RNF-3 moved homeScreen file in src package
