+++ src/navigation/HomeStack.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
-import HomeScreen from '../screens/HomeScreen';
+import HomeScreenTabStack from './HomeScreenTabStack';
 
 const Stack = createStackNavigator();
 
 export default function HomeStack() {
   return (
     <Stack.Navigator>
-      <Stack.Screen name='Home' component={HomeScreen} />
+      <Stack.Screen name='Home' component={HomeScreenTabStack} />
     </Stack.Navigator>
   );
 }
